+++ inbm/Changelog.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
  - (NEXMANAGE-837) Fix schema validation issue on failed UDM request sent from
  dispatcher to cloudadapter
  - (NEXMANAGE-900) Add UpdateFirmware to the common proto file
+ - 
+### Changed
+ - (NEXMANAGE-906) Truncate the granular log instead of removing it
 
 
 ## 4.2.6.1 - 2024-10-18